Update README.md
As part of SQL Advent Calendar where one needs to complete daily SQL challenges to light up the calendar!
This is a 24 day challenge started on 2 Dec 2024 and ending on 25 Dec 2024.
A new challenge will be unlocked every day at 12PM EST.
The questions are of random difficulty. Was fun solving them!